#15 fix: The remove of the song list (in the update button) was bad
Previously, when we did a remove of all the songs that we had in our list (to update them later) what we did was erase them (but in that case incorrectly, so that the elements that came out on the screen were completely random). Right now, with this commit, the update button already does this function correctly (without errors, becvause the delate now is correct).