#21 feat: Implemented the methods to eliminate a user in server
Several methods have been implemented so that a user can be erased, one for the object to be passed and the other for their user name to be passed.

It should be remembered that it is possible that in the future the two DAO methods will be unified, and that when the user needs something from the database, he will ask the server.