#21 fix: Solved different errors with the queries
Before in SQL queries, we have also made several errors. Now we do the queries with HQL.
Several queries have been adapted to this language and errors have been corrected.

At the moment it would be necessary to implement the method of deletion of the user and the call to an stored procedure.